requirements.txt
iniconfig==2.3.0
mss==10.1.0
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
pluggy==1.6.0
Pygments==2.19.2
pytesseract==0.3.13
pytest==9.0.2
python-dotenv==1.2.1
